# backend/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API для управления складом товаров с ролевой моделью доступа",
    "title": "MarketEase API",
    "contact": {},
    "version": "2.0"
  },
  "host": "localhost:8080",
  "basePath": "/api",
  "paths": {
...
    "/login": {
      "post": {
        "description": "Аутентификация пользователя по email и паролю",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["auth"],
        "summary": "Вход пользователя",
        "parameters": [
          {
            "description": "Данные для входа",
            "name": "credentials",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/controllers.LoginInput"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/controllers.TokenResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      }
    },
...
    "/refresh": {
      "post": {
        "description": "Получение нового токена доступа с помощью рефреш токена",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["auth"],
        "summary": "Обновление токена",
        "parameters": [
          {
            "description": "Рефреш токен",
            "name": "refresh_token",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/controllers.RefreshTokenInput"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/controllers.TokenResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      }
    },
    "/register": {
      "post": {
        "description": "Создать нового пользователя с ролью viewer",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["auth"],
        "summary": "Регистрация пользователя",
        "parameters": [
          {
            "description": "Данные пользователя для регистрации",
            "name": "user",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/controllers.RegisterInput"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/models.User"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "409": {
            "description": "Conflict",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "controllers.LoginInput": {
      "type": "object",
      "required": ["email", "password"],
      "properties": {
        "email": {
          "type": "string",
          "example": "john@example.com"
        },
        "password": {
          "type": "string",
          "example": "supersecret"
        }
      }
    },
    "controllers.RefreshTokenInput": {
      "type": "object",
      "required": ["refresh_token"],
      "properties": {
        "refresh_token": {
          "type": "string",
          "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        }
      }
    },
    "controllers.RegisterInput": {
      "type": "object",
      "required": ["email", "password", "username"],
      "properties": {
        "email": {
          "type": "string",
          "example": "john@example.com"
        },
        "password": {
          "type": "string",
          "minLength": 6,
          "example": "supersecret"
        },
        "username": {
          "type": "string",
          "example": "john_doe"
        }
      }
    },
    "controllers.TokenResponse": {
      "type": "object",
      "properties": {
        "access_token": {
          "type": "string",
          "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        },
        "refresh_token": {
          "type": "string",
          "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        },
        "token_type": {
          "type": "string",
          "example": "Bearer"
        }
      }
    },
...
    "models.ErrorResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Произошла ошибка"
        }
      }
    },
...
    "models.User": {
      "type": "object",
      "properties": {
        "createdAt": {
          "type": "string"
        },
        "deletedAt": {
          "type": "string"
        },
        "email": {
          "type": "string",
          "example": "john@example.com"
        },
        "id": {
          "type": "string",
          "example": "e7b1c2f1-567b-4b14-a8d7-cd5b08bfc9d0"
        },
        "role": {
          "type": "string",
          "example": "viewer"
        },
        "status": {
          "type": "string",
          "example": "active"
        },
        "updatedAt": {
          "type": "string"
        },
        "username": {
          "type": "string",
          "example": "john_doe"
        }
      }
    }
  },
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}